feat: move probe interval from scheduler config to manager config
