+++ manager/config/testdata/manager.yaml
@@ -88,3 +88,6 @@ metrics:
 
 network:
   enableIPv6: true
+
+probe:
+  interval: 30s
