+++ scheduler/service/service_v2.go
@@ -750,8 +750,7 @@ func (v *V2) SyncProbes(stream schedulerv2.Scheduler_SyncProbesServer) error {
 
 			logger.Infof("probe started: %#v", probedHosts)
 			if err := stream.Send(&schedulerv2.SyncProbesResponse{
-				Hosts:         probedHosts,
-				ProbeInterval: durationpb.New(v.config.NetworkTopology.Probe.Interval),
+				Hosts: probedHosts,
 			}); err != nil {
 				logger.Error(err)
 				return err
